requirements.txt
streamlit
requests
numpy
PyPDF2
python-docx
scikit-learn
